autopilot-supply-chain/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cryptography==42.0.7
redis==5.0.4
aio-pika==9.4.1
structlog==24.1.0
httpx==0.27.0
fastapi-mail==1.4.1
Jinja2==3.1.4
pyotp==2.9.0
authlib==1.3.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
prometheus-fastapi-instrumentator==6.1.0
websockets==12.0
python-dotenv==1.0.1
email-validator==2.1.1
